README.me
WIKIPEDIA: "In mathematics, the Euclidean algorithm, or Euclid's algorithm, 
is an efficient method for computing the greatest common divisor (GCD) of two numbers, 
the largest number that divides both of them without leaving a remainder. 
It is named after the ancient Greek mathematician Euclid, 
who first described it in his Elements (c. 300 BC)..."

This is an Example of the implementation of the Euclidian Algorithm